fix: rename ImageDetaiPageHeader to GalleryHeader
