+++ src/components/GalleryHeader/GalleryHeader.tsx
@@ -5,19 +5,14 @@ import { Avatar, Button } from '@/components';
 import { ReactComponent as ArrowIcon } from '@/assets/Button/arrow.svg';
 
 // types
-import { ImageDetailPageHeaderProps } from './ImageDetailPageHeader.type';
-import { Container, StyledImageDetailPageHeader } from './ImageDetailPageHeader.styled';
+import { GalleryHeaderProps } from './GalleryHeader.type';
+import { Container, StyledGalleryHeader } from './GalleryHeader.styled';
 
-export default function ImageDetailPageHeader({
-  title,
-  userName,
-  extraInfos,
-  to,
-}: ImageDetailPageHeaderProps): ReactElement {
+export default function GalleryHeader({ title, userName, extraInfos, to }: GalleryHeaderProps): ReactElement {
   return (
     <Container>
       <div className="title">
-        <StyledImageDetailPageHeader>{title}</StyledImageDetailPageHeader>;
+        <StyledGalleryHeader>{title}</StyledGalleryHeader>;
         <Button alt="Next Button" backgroundColor="blue" className="next" text="Next" img={ArrowIcon} />
       </div>
       <div className="content">
