+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as Avatar } from './Avatar/Avatar';
 export { default as Button } from './Button/Button';
 export { default as Image } from './Image/Image';
-export { default as ImageDetailPageHeader } from './ImageDetailPageHeader/ImageDetailPageHeader';
+export { default as GalleryHeader } from './GalleryHeader/GalleryHeader';
 export * from './Image/Image.type';
 export { default as Video } from './Video/Video';
 export { default as SearchBar } from './SearchBar/SearchBar';
